[Sher] divide pulbicData into small modules
