+++ src/pages/client/index.tsx
@@ -4,9 +4,7 @@ import { Divider, ConfigProvider } from 'antd';
 
 import { switchConfigProviderLocale } from '@/services/locale';
 import ChainSyncInfoTable from './component/ChainSyncInfoTable';
-import MinerInfoTable from './component/MinerInfoTable';
 import OperationBoard from './component/OperationBoard';
-import MiningInfoTable from './component/MiningInfoTable';
 
 
 const TableList: React.FC<{}> = () => {
@@ -19,9 +17,6 @@ const TableList: React.FC<{}> = () => {
         <OperationBoard />
         <Divider />
         <ChainSyncInfoTable />
-        <Divider />
-        <MinerInfoTable />
-        <MiningInfoTable />
       </ConfigProvider>
     </PageContainer >
   );
