+++ src/pages/client/locales/zh-CN.ts
@@ -9,18 +9,6 @@ export default {
     'strategy.name': '策略名称',
     'strategy.createAt': '创建时间',
 
-    'minerInfo.title': '矿工信息',
-    'minerInfo.stxAddress': 'STX地址',
-    'minerInfo.btcAddress': 'BTC地址',
-    'minerInfo.actualWins': '最终获胜数',
-    'minerInfo.totalWins': '燃烧链获胜数',
-    'minerInfo.totalMined': '总参与挖矿数',
-    'minerInfo.burn': '总燃烧量',
-
-    'miningInfo.title': '挖矿信息',
-    'miningInfo.stacksHeight': '链高度',
-    'miningInfo.burnfee': '燃烧量',
-
     'chainSyncInfo.title': '链同步信息',
     'chainSyncInfo.burn_block_height': '燃烧链高度',
     'chainSyncInfo.stacks_tip_height': '主链高度',
