+++ src/pages/publicData/component/MinerInfoTable.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { MinerInfo, MinerInfoQueryParams } from "@/services/client/data";
+import { MinerInfo, MinerInfoQueryParams } from "@/services/publicdata/data";
 import ProTable, { ProColumns } from "@ant-design/pro-table";
 import { FormattedMessage, useModel } from "umi";
 
 const MinerInfoTable: React.FC<{}> = () => {
     const { operationBoardState } = useModel('client.operationBoard');
-    const { minerAddress } = operationBoardState;
-    const { queryMinerInfo } = useModel('client.minerInfo');
+    let { minerAddress } = operationBoardState;
+    const { queryMinerInfo } = useModel('publicData.minerInfo');
 
     const minerInfoColumns: ProColumns<MinerInfo>[] = [
         {
