+++ src/pages/publicData/models/minerInfo.ts
@@ -1,5 +1,5 @@
-import { getMinerInfo } from "@/services/client/Client";
-import { MinerInfo, MinerInfoQueryParams } from "@/services/client/data";
+import { getMinerInfo } from "@/services/publicdata/miningInfo";
+import { MinerInfo, MinerInfoQueryParams } from "@/services/publicdata/data";
 import { useState } from "react";
 
 export interface MinerInfoState {
